push first

+++ Membership/API/Members/Payments/IPaymentManager.vb
@@ -2,10 +2,10 @@
     'Probably should cache the result from this.
     Function getUnpaidCharges(member As IMember) As HashSet(Of IMemberCharge)
 
-    Function addCharge(member As IMember, timestamp As Date, desc As String, amount As Double) As IMemberCharge
+    Function addCharge(member As IMember, timestamp As Date, desc As String, amount As Decimal) As IMemberCharge
 
     Function invoice(charges As HashSet(Of IMemberCharge)) As IPaymentInvoice
-    Function pay(invoice As IPaymentInvoice, amount As Double) As IMemberPayment
+    Function pay(invoice As IPaymentInvoice, amount As Decimal) As IMemberPayment
 
     Function removePayment(payment As IMemberPayment) As Boolean
     ''' <summary>
